requirements.txt
absl-py==2.3.1
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.60.1
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
mediapipe==0.10.5
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1732314388626/work/dist/numpy-2.0.2-cp39-cp39-macosx_11_0_arm64.whl#sha256=f5ae29dd6f735c6dcad61765cffcfb32661c04e515c759b21399c7a3a39430a4
opencv-contrib-python==4.12.0.88
opencv-python==4.10.0
opencv-python-headless==4.10.0
packaging==25.0
pillow==11.3.0
playsound==1.3.0
protobuf==3.20.3
pycparser==2.23
pygame==2.6.1
pyparsing==3.2.5
pyserial @ file:///home/conda/feedstock_root/build_artifacts/pyserial_1734140157046/work
python-dateutil==2.9.0.post0
six==1.17.0
sounddevice==0.5.3
zipp==3.23.0
